backend/requirements.txt
Flask==2.3.3
Flask-Cors==4.0.0
Flask-JWT-Extended==4.5.3
Flask-SQLAlchemy==3.1.1
Flask-Mail==0.9.1
PyMySQL==1.1.0
python-dotenv==1.0.0
bcrypt==4.0.1
cryptography==41.0.4
itsdangerous==2.1.2
requests==2.31.0
email-validator==2.0.0

google-auth==2.23.4
google-auth-oauthlib==1.0.0
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0

vosk==0.3.45
pydub==0.25.1
openai==0.28.0
websockets==11.0.3p